Redesigned the start/home screen
Because:
- There should be two instead of four buttons for the navigation

Added a description next to the buttons to guide the user.
