+++ app/src/main/java/com/example/vpmanager/createStudy/createStudyFragment_Base.java
@@ -1,4 +1,4 @@
-package com.example.vpmanager.views;
+package com.example.vpmanager.createStudy;
 
 import android.os.Bundle;
 
@@ -10,13 +10,14 @@ import android.view.ViewGroup;
 
 import com.example.vpmanager.R;
 
-public class createStudyFragment_StepOne_Contact extends Fragment {
+public class createStudyFragment_Base extends Fragment {
+
 
     //Parameter:
     //Return values:
     //Sets the current fragment for the activity
-    public createStudyFragment_StepOne_Contact() {
-        createStudyActivity.currentFragment = 2;
+    public createStudyFragment_Base() {
+        createStudyActivity.currentFragment = 0;
     }
 
     @Override
@@ -27,11 +28,12 @@ public class createStudyFragment_StepOne_Contact extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_create_study_step_one_contact, container, false);
+        return inflater.inflate(R.layout.fragment_create_study_base, container, false);
     }
 
     @Override
     public void onViewCreated(View view, Bundle savedInstanceState) {
+
     }
 
 }
