+++ app/src/main/java/com/example/vpmanager/createStudy/createStudyFragment_StepFive.java
@@ -1,4 +1,4 @@
-package com.example.vpmanager.views;
+package com.example.vpmanager.createStudy;
 
 import android.app.DatePickerDialog;
 import android.app.TimePickerDialog;
