+++ app/src/main/java/com/example/vpmanager/createStudy/createStudyFragment_StepFour_Presence.java
@@ -1,23 +1,22 @@
-package com.example.vpmanager.views;
+package com.example.vpmanager.createStudy;
 
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import androidx.fragment.app.Fragment;
+
 import com.example.vpmanager.R;
 
-public class createStudyFragment_Base extends Fragment {
+public class createStudyFragment_StepFour_Presence extends Fragment {
 
 
     //Parameter:
     //Return values:
     //Sets the current fragment for the activity
-    public createStudyFragment_Base() {
-        createStudyActivity.currentFragment = 0;
+    public createStudyFragment_StepFour_Presence() {
+        createStudyActivity.currentFragment = 5;
     }
 
     @Override
@@ -28,12 +27,10 @@ public class createStudyFragment_Base extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_create_study_base, container, false);
+        return inflater.inflate(R.layout.fragment_create_study_step_four_presence, container, false);
     }
 
     @Override
     public void onViewCreated(View view, Bundle savedInstanceState) {
-
     }
-
 }
