+++ app/src/main/res/layout/activity_create_study.xml
@@ -5,7 +5,7 @@
     android:id="@+id/drawerLayoutCreate"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".views.createStudyActivity">
+    tools:context=".createStudy.createStudyActivity">
 
     <androidx.coordinatorlayout.widget.CoordinatorLayout
         android:layout_width="match_parent"
@@ -45,11 +45,11 @@
                 android:layout_weight="8"
                 android:background="?android:windowBackground"
                 app:cardCornerRadius="20dp"
-                tools:context=".views.createStudyFragment_Base">
+                tools:context=".createStudy.createStudyFragment_Base">
 
                 <fragment
                     android:id="@+id/fragment_container"
-                    android:name="com.example.vpmanager.views.createStudyFragment_Base"
+                    android:name="com.example.vpmanager.createStudy.createStudyFragment_Base"
                     android:layout_width="match_parent"
                     android:layout_height="match_parent"
                     android:layout_marginLeft="10dp"
