+++ app/src/main/res/layout/fragment_create_study_base.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?android:windowBackground"
-    tools:context=".views.createStudyFragment_Base">
+    tools:context=".createStudy.createStudyFragment_Base">
 
     <LinearLayout
         android:layout_width="match_parent"
