+++ app/src/main/res/layout/fragment_create_study_final_step_three.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".views.createStudyFragment_finalStep_three"
+    tools:context=".createStudy.createStudyFragment_finalStep_three"
     android:background="?android:windowBackground">
