+++ app/src/main/res/layout/fragment_create_study_final_step_two.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".views.createStudyFragment_finalStep_two"
+    tools:context=".createStudy.createStudyFragment_finalStep_two"
     android:background="?android:windowBackground">
