+++ app/src/main/res/layout/fragment_create_study_step_five.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?android:windowBackground"
-    tools:context=".views.createStudyFragment_StepFive">
+    tools:context=".createStudy.createStudyFragment_StepFive">
 
 
     <LinearLayout
