+++ app/src/main/res/layout/fragment_create_study_step_one.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?android:windowBackground"
-    tools:context=".views.createStudyFragment_StepOne">
+    tools:context=".createStudy.createStudyFragment_StepOne">
 
 
     <LinearLayout
