+++ app/src/main/res/layout/fragment_create_study_step_two.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".views.createStudyFragment_StepTwo"
+    tools:context=".createStudy.createStudyFragment_StepTwo"
     android:background="?android:windowBackground">
