+++ app/src/main/res/navigation/nav_graph_main.xml
@@ -66,7 +66,7 @@
     </fragment>
     <activity
         android:id="@+id/createStudyActivity"
-        android:name="com.example.vpmanager.views.createStudyActivity"
+        android:name="com.example.vpmanager.createStudy.createStudyActivity"
         android:label="Studien erstellen"
         tools:layout="@layout/activity_create_study" />
     <fragment
